Add files via upload
In step 1 our program compares the average price of a category from a specific season of two different brands. The program includes a UI where the user is shown a dropdown menu to choose the category, brand, and season they desire. This is useful for both the customer and retail manager for quick insights on price transparency and availability of comparing the price with another brand without the need of looking it up on the website. Step 1 ends with a loop, if the user has all the necessary information the code ends and if not the UI appears again for the user to choose again.