Add files via upload
In step 2 the program first asks the user once again to choose a category and season, after this the program created a bar chart showcasing all brands with their respective average price of that certain category (product type) on the season desired. This is very helpful again for both customers and retail stores managers or analysts for a more broad concept of pricing and at the same time for the customer, if their decision is based on price, to choose what to buy depending on it being the most expensive or the cheapest.